+++ app/api/entities/events.ts
@@ -1,0 +1,1 @@
+export const EntitySavedEvent = 'entity:saved';
