+++ app/api/entities/specs/entities.spec.js
@@ -183,56 +183,6 @@ describe('entities', () => {
         .catch(catchErrors(done));
     });
 
-    it('should update ix suggestions on entity update, if relevant metadata is changed', async () => {
-      const updateSpy = jest.spyOn(Suggestions, 'updateStates');
-
-      const doc = {
-        _id: batmanFinishesId,
-        sharedId: 'shared',
-        title: 'Batman finishes in other words',
-      };
-      const user = { _id: db.id() };
-      let saved = await entities.save(doc, { user, language: 'en' });
-      expect(updateSpy).not.toHaveBeenCalled();
-
-      let toSave = {
-        ...doc,
-        metadata: {
-          ...saved.metadata,
-          text: [{ value: 'new text value' }],
-          property1: [{ value: 'new property 1 value' }],
-        },
-      };
-      saved = await entities.save(toSave, { user, language: 'en' });
-      expect(updateSpy).not.toHaveBeenCalled();
-
-      toSave = {
-        ...doc,
-        metadata: {
-          ...saved.metadata,
-          property2: [{ value: 'new property 2 value' }],
-        },
-      };
-      saved = await entities.save(toSave, { user, language: 'en' });
-      expect(updateSpy).toHaveBeenCalledWith({ entityId: saved.sharedId });
-
-      updateSpy.mockClear();
-      toSave = {
-        ...doc,
-        metadata: {
-          ...saved.metadata,
-          text: [{ value: 'second new text value' }],
-          property1: [{ value: 'second new property 1 value' }],
-          property2: [{ value: 'second new property 2 value' }],
-          description: [{ value: 'new description value' }],
-        },
-      };
-      saved = await entities.save(toSave, { user, language: 'en' });
-      expect(updateSpy).toHaveBeenCalledWith({ entityId: saved.sharedId });
-
-      updateSpy.mockRestore();
-    });
-
     it('should index the newly created documents', done => {
       const doc = { title: 'the dark knight', template: templateId };
       const user = { _id: db.id() };
