+++ app/api/eventListeners.ts
@@ -1,0 +1,6 @@
+import { EventsBus } from './eventsbus';
+import { Suggestions } from './suggestions/suggestions';
+
+export const registerEventListeners = (eventsBus: EventsBus) => {
+  Suggestions.registerEventListeners(eventsBus);
+};
