+++ app/api/eventsbus/EventsBus.ts
@@ -1,0 +1,5 @@
+import { EventEmitter } from 'events';
+
+export class EventsBus extends EventEmitter {}
+
+export const applicationEventsBus = new EventsBus();
