+++ app/api/eventsbus/index.ts
@@ -1,0 +1,1 @@
+export { EventsBus, applicationEventsBus } from './EventsBus';
